Move env contextmanager to base testcase class.
